Add simple container listing
To run this, simply upload everything under app/ to Swift and load the
index.html page in your browser. The page will query Swift for the
containers.

This assumes that you've already logged in and somehow have access to
Swift. This is the case when using zebra.zerovm.org, but for a
devstack installation of Swift you'll need to configure staticweb and
setup read access for anonymous users yourself.