e2e: add end-to-end test for root page
This works by injecting JavaScript code into the running browser to
configure the $httpBackend service.

This is done using the addMockModule protractor method which installs
code to run *before* Angular has been loaded. The code is run before
every browser.get call.

The page much include angular-mocks.js for this to work. This script
is added to the page before each test is run. If multiple tests are
run the file would be added multiple times since the browser doesn't
do a full page refresh between the tests.