e2e: test sorting container listing by count and size
For the size column, we cannot use the by.repeater(...) helper locator
since the column method doesn't work when there's no content to match
on. We select the column by CSS instead then -- this will of course
break if the columns are ever reordered, but the breakage will be easy
fix then.