e2e: move $httpBackend configuration inside SwiftSimulator
This moves more boiler-plate code away from the tests and into the
SwiftSimulator. This requires Protractor version 0.23.0 or later,
we're now using 1.0.0.

The special arguments array is a local variable automatically
available within all functions. We must bind it to a local variable to
access in the nested function, otherwise the inner arguments variable
would shadow the outer.