e2e: update protractor version to ~1.0.0
This is to get support for the third argument in addMockModule. In

  https://github.com/angular/protractor/commit/b7afa8791ba9

support was introduced for passing arguments to the initialization
code for the mock module. We will use this to pass in the containers
for the Swift simulator -- today we do this by actually hard-coding
the container list inside the mock module initialization function.

Remember run 'npm update protractor' to install the new version.

The new version of protractor will warn if a locator selected more
than one element. This is why we need to use a more precise CSS
expression when selecting the first table header.