eslint: add config file
This file whitelists the globals defined by Karma, Jasmine, and
Protractor. This is to avoid having them show up as undefined
variables.

The next version of ESLint will have an environment setting for adding
the Jasmine globals:

  https://github.com/eslint/eslint/commit/e3e55bac89dc

Until that is released, we'll just the define the globals here.

This config file also disables the no-comma-dangle rule. Dangling
commas cause trouble in IE8, but we're not running our tests with that
browser.