swift: make SwiftClient an Angular service
This means that the Angular injector will create the SwiftClient for
us when we inject the $swift service into a controller or other
function. See the Angular documentation:

  https://docs.angularjs.org/guide/providers#service-recipe

The client will now only be instantianted once instead of being
instantiated per controller. Before it was instantiated by the
controllers, and they are in turn instantiated on every route change.
The hope is that keeping a single SwiftClient around will make it
easier to login just once.