controllers: avoid iterating over items using length
Iterating using explicit indexes and a for loop will not work if
elements has been deleted from the array. We will have such elements
when we begin deleting objects.

The use of of Array.prototype.forEach, Array.prototype.some, and
Array.prototype.every implies a requirements for IE9 and later. I'll
add a note to the README to make this explicit.