filters: add notUndefined filter
We will use this to filter out the deleted objects from our listings.
Deleting an array entry in JavaScript does not remove it, instead it
is set to the the value "undefined". When iterating over an array with
ngRepeat, these values are still shown, so we need to filter them.