grunt: move Travis ESLint configuration to Grunt (#21)
Install the grunt-cli somewhere global on your system. This gives you
the "grunt" command, which will use the version of Grunt installed
locally for your project.

Then run "grunt eslint" to start eslint with the same options as we
use on Travis.

The explicit dependency on eslint is no longer needed -- the
grunt-eslint module has its own dependency on eslint.