grunt: add protractor task (#21)
The grunt-protractor-runner task depends on protractor, so we no
longer need to depend on it explicitly.

The latest version of grunt-protractor-runner is 1.1.4 at the time of
writing. That version moved the installation of the webdriver from a
postinstall script to a pretest script. Without the

  webdriver-manager update

call, one cannot run protractor. With grunt-protractor-runner, the
webdriver-manager script resides in

  node_modules/grunt-protractor-runner/node_modules/.bin/webdriver-manager update

which is a quite unwieldy path. So instead of referencing it from our
package.json file, I opted to use grunt-protractor-runner 1.1.3 which
knows how to run webdriver-manager automatically.

The update-webdriver and protractor npm scripts are now obsolete, the
latter has been replaced by 'grunt protractor'.