package: make protractor a top-level dependency
In a later commit, we will be using grunt-protractor-coverage to
collect coverage data from the protractor tests. That module comes
with a protractor_coverage task which calls protractor -- but without
declaring the dependency.

So we need to make protractor a top-level dependency so that
grunt-protractor-coverage can see it. However, that breaks
grunt-protractor-runner version 1.1.3, invokes webdriver-manager in
its postinstall script. The invocation fails with:

  $ npm install
  -
  > grunt-protractor-runner@1.1.3 postinstall /home/mg/tmp/node_modules/grunt-protractor-runner
  > node scripts/webdriver-manager-update

  webdriver-manager path: /home/mg/tmp/node_modules/protractor/bin/webdriver-manager

  module.js:340
      throw err;
            ^
  Error: Cannot find module 'adm-zip'

Upgrading to grunt-protractor-runner 1.1.4 or later fixes this since
it no longer tries to invoke webdriver-manager. We now invoke it
ourselves before running the Travis tests.