grunt: add start task
This replaces the "npm start" script. I was having problems with a
template file having been sent with a max-age=3600 header, resulting
in my seeing the old version after reloading.

Using a more full-featured static web server will likely be useful in
the future. For one, we can begin serving the bower_components
directory with an expires header. We can also consider serving the app
from a more Swift-like URL.