container: add single-file upload (fixes #42)
This uses the HTML 5 file API to read the local file before sending it
to the server with a PUT request. As such, this only works on IE 10+
and other modern browsers:

  http://caniuse.com/#feat=fileapi

To support older browsers, we would have to use the Form POST
Middleware for Swift, please see #62.

The UI is very basic, see #66 about using a button with better
styling.