swift-sim: make SwiftSimultor an Angular service
This makes the simulator a singleton for the test run. Instead of
repeatedly configuring $httpBackend, the backend is now configured
once at the test start and the response functions consult the current
this.containers and this.objects as needed to compute their responses.

This makes it much easier to examine the state of the simulator at
runtime since the state is visible instead of being hidden inside
$httpBackend.