swift-mock: clear mock modules
Modules registered with addMockModule stay for the entire duration of
the test run. So when each test adds more and more mock modules, we
end up executing the code again and again.

We now clear the existing list of mock modules when the swift mock is
injected. The mock is then only loaded once. Furthermore, containers
and objects are then also only configured once.