swift: use $upload.http for uploading files
This is to get support for upload progress. The ng-file-upload package
will monkeypatch the XMLHttpRequest a little so that it can get access
to the XHR created by $http. It needs access to it so it can attach an
onprogress event handler.

We're not currently using anything else from ng-file-upload.