swift: pass extra headers to auth promise
This allows $swift to tell $auth about extra headers that should be
included when pending requests are retried.

Before, while the X-Auth-Token header was correctly added to
$swift._headers, it was not included when $auth retried the pending
requests -- with the result that the failed request would trigger
authentication again. For the cookie-based authentication in liteauth
this didn't matter (since the browser would send the cookie when the
request was retried), but it matters when we authenticate against
Keystone where we need to include the auth token with every request.

An alternative solution would be for $swift to update $http.defaults.
However, that would be changing global state and obscure the flow
between the two components.