swift: let $http add query string parameters
This changed the URL used in a test because the internal buildUrl
function used in $http sorts the parameters before adding them to the
URL. In other words, we can expect the new output to be stable.

We also had to change the regexps used in the Swift similator so that
they can handle the lack of '?' when listing a container.