app: introduce object state with ObjectCtrl
This splits the ContainerCtrl into two, one controller for handling
listing pseudo-directories (paths ending with a trailing slash) and
another for listing single objects (other paths).

Before, drilling down in the hiearchy of pseudo-directories would
eventually take you to a page with a "listing" of a single object.
This happened when the prefix excluded all other objects. This page is
now empty -- we will use it to show object specific metadata.