e2e: add timezone to last modified dates
This makes it unnecessary to convert the dates to UTC afterwards.

Infact, the conversion was wrong since Date.getTimezoneOffset returns
the local timezone offset *regardles* of the timezone used to create
the timestamp. So a Date created from "2014-08-16T13:33:21.000Z" would
still report a non-zero timezone offset (unless you run the test in
the GMT timezone).

We now only adjust for the timezone offset when we set the last
modified date to the current time. This is okay since we're dealing
with a new Date that we have created ourselves.