swift-sim: change internal representation of objects
Before, objects were represented by the JSON that we needed to return
in listContainer.

This was limiting since we will want to store more information per
object than what this method returns. The first example of extra data
will be custom metadata headers.

We therefore now represent the objects by the headrers we need to
return when headObject is called.