swift-sim: remove setContainers method (fixes #92)
Before, containers and objects were stored in separate variables. They
are now merged into a single data object. The keys of this object are
the containers. Each container has an 'objects' property which maps
object names to the objects themselves.

This simplifies the code a lot by letting listContainers compute its
result dynamically based on the objects registered. This will also
allow us to add container metadata since we now have an object to
attach this data to.