object: forbid removing the Content-Type header
Swift behaves in a peculiar way with regards to the Content-Type
header. While other headers can be unset, the Content-Type header is
always present. Trying to unset it will leave it unchanged.

This is now reflected in the simulator which will always copy the old
Content-Type header in postObject. In the UI, the trash can icon has
been removed for that header.