edit-modal: support editing object content (fixes #43)
Testing this required a change in how the simulator stores the object
content. Before, putObject would simply store the File object in
object.content, it now reads the File and stores the text. This is
needed since getObject returns object.content, and Angular does not
turn this into a string for us.