controllers: extract EditModalCtrl
This collects the logic for the edit-modal in a single controller.
Doing so simplifies the ObjectCtrl and has the benefit that the
controller can operate directly on the scope used by the modal.

Before, the modal was using a child scope of the scope that we
configured -- this meant that forms and fields published on the scope
by ngModel and ngForm weren't available to us.