swift-mock: set async = false on script tags
In HTML5, dynamically script elements default to async = false,
meaning that they'll be downloaded in parallel and executed in an
unknown order.

This is most likely the reason why we've seen random errors like:

  Error: Error while running module script swiftBrowserE2E:
  [$injector:nomod] Module 'swiftBrowserE2E' is not available!

in the Travis output. An example is this build:

  https://travis-ci.org/zerovm/swift-browser/builds/39145010

Explicitly setting the async attribute to false makes the browser
execute the scripts in the order they are added. See

  http://www.html5rocks.com/en/tutorials/speed/script-loading/

for more background information.