swift-sim: consistently use lowercase for HTTP headers
We now make sure to only use lowercase header names internally. This
fixes subtle problems and will let us simplify some tests in the
future now that we know the case of the returned headers.

One problem fixed directly by this was how we counted the bytes in a
container: putObject would set 'content-lenth', but listContainers
used 'Content-Length' instead.