eslint: forbid unused function arguments
We added an exception to this way back with the idea that this would
make callbacks nicer since they could all have the same signature.
However, this also meant that we did not detect when arguments became
unused due to refactoring. As an example, ObjectCtrl had two unused
injected arguments.