grunt: add mock task (fixes #117)
This generates a mock.html which will load the Swift simulator and
configure a basic test environment. The simulator is loaded in the
same way Protractor loads it (this can probably be simplified now).

You will be working on an in-memory database which reflects the
changes you make until you reload the page. Navigating using the
in-page links is okay since that doesn't trigger a full page reload.