container: load objects incrementally (fixes #159, fixes #160)
Loading the entire container listing with a single GET makes the UI
freeze when the objects are inserted in the table. We now load the
object names in small chunks instead and this gives a much more
responsive application. The time to seeing the first object is also
much improved since we can quickly insert the first chunk.

This also ensures that we can handle containers with more than 10k
objects. Swift will be default not return more than 10k results for a
single GET, and the code can now handle this.