container: load objects in bigger and bigger chunks
This saves some round-trips to the server. By starting with a small
limit of 8, we get a quick response and can then fire off a query for
the next chunk of 16 objects.

We limit the chunk size to 128 since this seems to be the size where
the transfer and processing time becomes a problem again.