upload: fix type error
The problem is that parseInt takes a string, not a number. So when we
passed a number this was first cast to a string, which was then parsed
as an integer, effectively throwing away the decimal part. That is
silly when Angular has a filter that can round then number for us.

This was detected with flow: http://flowtype.org/.