package: use protractor 1.4.0
This allows us to use the new directConnect option in Protractor. This
will connect directly to the browser instead of going through
Selelium. In testing locally, this seems to be 2-5 sec faster.

This also means that we no longer need to download the webdrivers. In
the Travis build, this takes 7-8 sec, which we can now save.